+++ src/django_dumpdata_one/management/commands/dumpdata_one.py
@@ -1,5 +1,6 @@
 import json
 from django.apps import apps
+from django.core.files import File
 from django.core.management.base import BaseCommand
 from django.core.serializers.json import DjangoJSONEncoder
 
@@ -12,6 +13,8 @@ class Command(BaseCommand):
         parser.add_argument('--order', type=str, default='pk')
         parser.add_argument('--fields', type=str)
         parser.add_argument('--filter', type=str)
+        parser.add_argument('--full_url', type=str)
+        parser.add_argument('--limit', type=int)
 
     def handle(self, *args, **options):
 
@@ -19,6 +22,8 @@ class Command(BaseCommand):
         fields = options['fields']
         order = options['order']
         filters = self.get_filter_dict(options['filter'])
+        full_urls = options['full_url']
+        limit = options['limit']
 
         Model = apps.get_model(app_model)
 
@@ -30,30 +35,43 @@ class Command(BaseCommand):
         if 'pk' not in fields:
             fields.append('pk')
 
+        full_urls = full_urls.split(',') if full_urls is not None else []
+
         records = Model.objects.all()
         if filters:
             records = records.filter(**filters)
 
         order_by = self.get_order_list(order)
 
-        records = records.order_by(*order_by).values(*fields)
+        records = records.order_by(*order_by)
+        if limit:
+            records = records[:limit]
 
-        dump_structure = self.get_dump_structure(app_model, records)
+        dump_structure = self.get_dump_structure(
+            app_model, records, fields, full_urls
+        )
         result = json.dumps(dump_structure, cls=DjangoJSONEncoder)
         return result
 
-    def get_dump_structure(self, app_model, records):
+    def get_dump_structure(self, app_model, records, fields, full_urls):
         results = []
         for item in records:
-            fields = {}
-            for key, value in item.items():
-                if key != "pk":
-                    fields[key] = value
+            values = {}
+            for field in fields:
+                if field != "pk":
+                    model_field = getattr(item, field)
+                    if isinstance(model_field, File):
+                        values[field] = model_field.name
+                    else:
+                        values[field] = model_field
+
+                if field in full_urls:
+                    values[f"{field}__full_url"] = getattr(item, field).url
 
             item_structure = {
                 "model": app_model,
-                "pk": item.get("pk"),
-                "fields": fields
+                "pk": getattr(item, "pk"),
+                "fields": values
             }
             results.append(item_structure)