+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## [0.8.2] - 2020-07-01
+### Added
+- dump full file/image url `--full_url={field_name},{other_file_name}`
+- limit number of records `--limit={number}`
+
+### Changed
+- rewrite way to retrieve a field value
+
 ## [0.8.1] - 2020-06-02
 ### Added
 - dump all model fields with argument `--fields=*`